+++ python/pathway/io/__init__.py
@@ -33,7 +33,12 @@ from pathway.io import (
     slack,
     sqlite,
 )
-from pathway.io._subscribe import OnChangeCallback, OnFinishCallback, subscribe
+from pathway.io._subscribe import (
+    OnChangeCallback,
+    OnChangeCallbackAsync,
+    OnFinishCallback,
+    subscribe,
+)
 from pathway.io._synchronization import (
     SynchronizedColumn,
     register_input_synchronization_group,
@@ -59,6 +64,7 @@ __all__ = [
     "pyfilesystem",
     "python",
     "OnChangeCallback",
+    "OnChangeCallbackAsync",
     "OnFinishCallback",
     "redpanda",
     "slack",